requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.7
django-celery-beat==2.5.0
django-ckeditor-5==0.2.10
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-js-asset==2.1.0
django-mptt==0.15.0
django-recaptcha==3.0.0
django-taggit==5.0.1
django-timezone-field==6.0.1
kombu==5.3.3
mypy==1.6.1
mypy-extensions==1.0.0
Pillow==10.1.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
python-crontab==3.0.0
python-dateutil==2.8.2
pytils==0.4.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.9
